feat: refrain from re-creating marks
Previously all extmarks were re-created everytime a buffer is entered.
This causes issues when changes are made to files, because the mark
position is difficult to correctly determine when changes are made to
files.
Now only removed comment marks are removed, and new comments created.
This leverages the built-in capability of marks to move along with
changes in the code.
The only case where it will not work is when re-loading the plugin after
the changes were done, (without committing / pushing the changes to
the PR.)